Update XDR for history ledger to include base percentage fee

+++ services/horizon/internal/db2/schema/migrations/11_add_base_percentage_fee_to_history_ledger.sql
@@ -1,3 +1,3 @@
 -- Precision 5, scale 4 to handle this number
 -- 0.0045, which represents a percentage fee of 0.45%.
-ALTER TABLE history_ledgers ADD base_percentage_fee NUMERIC(5, 4);
+ALTER TABLE history_ledgers ADD base_percentage_fee INTEGER;
